00._Mandatories/01._Mandatory_I/01a-data-parsing/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = ""
authors = [
    {name = "steffengroenandersen",email = "steffengroenandersen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "xmltodict (>=0.14.2,<0.15.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
